Removed copy-webpack-plugin
The copy-webpack-plugin was added in the initial commit to copy the values.json from the public folder.  Instead the values.json file is imported into the code but webpack is configured to return the filename instead of the data for the file so that we can call fetch with the filename.